Use `VecBuffer` instead of `GenericImage`.
Added some more benchmarks (we may cherry-pick them on master ?).

```
before
test contrast::test::bench_threshold                                    ... bench:   1,175,968 ns/iter (+/- 12,929)
test contrast::test::bench_threshold_mut                                ... bench:     341,623 ns/iter (+/- 4,349)
test contrast::test::bench_equalize_histogram                           ... bench:   2,227,185 ns/iter (+/- 20,862)
test contrast::test::bench_equalize_histogram_mut                       ... bench:   1,432,186 ns/iter (+/- 33,085)

after
test contrast::test::bench_threshold                                    ... bench:      31,795 ns/iter (+/- 437)
test contrast::test::bench_threshold_mut                                ... bench:      12,227 ns/iter (+/- 274)
test contrast::test::bench_equalize_histogram                           ... bench:     948,128 ns/iter (+/- 11,067)
test contrast::test::bench_equalize_histogram_mut                       ... bench:     929,868 ns/iter (+/- 12,838)
```