@theotherphil 2016-01-17
... and much shorter code!

Only iter(), iter_mut() and clone() required for our hypothetical new image trait so far. The one slight worry I have is that the GenericImage already has a pixels_mut method, but it's marked as deprecated with some comments about lifetime issues and this not being possible in safe rust. We'll have to check whether the same issues cause problems for us.