@theotherphil 2016-01-17
I've not read it yet, but this thread discusses the problem: http://stackoverflow.com/questions/25730586/returning-mutable-references-from-an-iterator/25748645#25748645